recreate web socket connection on didCompleteWithError
