+++ AmplifyPlugins/API/Sources/AWSAPIPlugin/AppSyncRealTimeClient/AppSyncRealTimeClient+HandleRequest.swift
@@ -23,16 +23,6 @@ extension AppSyncRealTimeClient {
         _ request: AppSyncRealTimeRequest,
         timeout: TimeInterval = 5
     ) async throws {
-        // Perform auth token decoration before starting the response timeout.
-        // Auth token retrieval can take a variable amount of time, especially
-        // under contention with multiple concurrent subscriptions. The timeout
-        // should only measure the server round-trip after the write, not local
-        // preparation work.
-        let decoratedRequest = await self.requestInterceptor.interceptRequest(
-            event: request,
-            url: self.endpoint
-        )
-
         var responseSubscriptions = Set<AnyCancellable>()
         try await withCheckedThrowingContinuation { [weak self] (continuation: CheckedContinuation<Void, Swift.Error>) in
             guard let self else {
@@ -41,7 +31,7 @@ extension AppSyncRealTimeClient {
                 return
             }
 
-            // listen to response — timeout starts now, after auth decoration
+            // listen to response
             subject
                 .setFailureType(to: AppSyncRealTimeRequest.Error.self)
                 .flatMap { Self.filterResponse(request: request, result: $0) }
@@ -57,9 +47,13 @@ extension AppSyncRealTimeClient {
                 }, receiveValue: { _ in })
                 .store(in: &responseSubscriptions)
 
-            // send the already-decorated request
+            // sending request
             Task {
                 do {
+                    let decoratedRequest = await self.requestInterceptor.interceptRequest(
+                        event: request,
+                        url: self.endpoint
+                    )
                     let requestJSON = try String(data: Self.jsonEncoder.encode(decoratedRequest), encoding: .utf8)
                     try await self.webSocketClient.write(message: requestJSON!)
                 } catch {
