+++ AmplifyPlugins/API/Tests/AWSAPIPluginTests/AppSyncRealTimeClient/AppSyncRealTimeClientWriteFailureTests.swift
@@ -16,8 +16,8 @@ import XCTest
 /// After iOS suspends the app process, the kernel defuncts TCP flows bound to
 /// the WebSocket connection. When the app returns to foreground and attempts to
 /// resubscribe, `WebSocketClient.write()` throws NSPOSIXErrorDomain Code=53
-/// ("Software caused connection abort"). The library must detect this write
-/// failure and trigger a full reconnection cycle so subscriptions can recover.
+/// ("Software caused connection abort"). The library must propagate this error
+/// to subscribers so they don't remain stuck in "subscribing" state.
 class AppSyncRealTimeClientWriteFailureTests: XCTestCase {
 
     /// Verifies that when a WebSocket write fails with a connection error,
@@ -73,329 +73,6 @@ class AppSyncRealTimeClientWriteFailureTests: XCTestCase {
         withExtendedLifetime(cancellables) { }
     }
 
-    /// Verifies that when a connected AppSyncRealTimeClient receives a
-    /// `.disconnected` event (as would happen when the WebSocket detects a
-    /// dead connection), it transitions to `.connectionDropped` state and
-    /// triggers reconnection when the WebSocket comes back online.
-    ///
-    /// - Given:
-    ///    - An AppSyncRealTimeClient in connected state with an active subscription.
-    /// - When:
-    ///    - The WebSocket emits a `.disconnected` event (simulating socket death
-    ///      detection after process suspension).
-    ///    - Followed by a `.connected` event (simulating successful reconnection).
-    /// - Then:
-    ///    - The client sends `connection_init` again (reconnection).
-    ///    - After receiving `connection_ack`, it resumes existing subscriptions
-    ///      by sending a new `start` request.
-    func testConnectionDrop_withActiveSubscription_shouldReconnectAndResubscribe() async throws {
-        var cancellables = Set<AnyCancellable>()
-        let mockWebSocketClient = MockWebSocketClient()
-        let mockAppSyncRequestInterceptor = MockAppSyncRequestInterceptor()
-        let appSyncClient = AppSyncRealTimeClient(
-            endpoint: URL(string: "https://example.com")!,
-            requestInterceptor: mockAppSyncRequestInterceptor,
-            webSocketClient: mockWebSocketClient
-        )
-
-        let id = UUID().uuidString
-
-        let initialSubscribed = expectation(description: "Initial subscription established")
-        let resubscribedAfterDrop = expectation(description: "Subscription re-established after connection drop")
-        resubscribedAfterDrop.assertForOverFulfill = false
-
-        let subscribedCount = AtomicCounter()
-
-        await mockWebSocketClient.setStateToConnected()
-
-        // Wire up the mock to respond to protocol messages
-        await mockWebSocketClient.actionSubject
-            .sink { action in
-                switch action {
-                case .connect:
-                    Task {
-                        await mockWebSocketClient.subject.send(.connected)
-                    }
-                case .write(let message):
-                    guard let response = try? JSONDecoder().decode(
-                        JSONValue.self,
-                        from: message.data(using: .utf8)!
-                    ) else { return }
-
-                    switch response.type?.stringValue {
-                    case "connection_init":
-                        Task {
-                            try await Task.sleep(nanoseconds: 50_000_000)
-                            await mockWebSocketClient.subject.send(.string("""
-                                {"type": "connection_ack", "payload": { "connectionTimeoutMs": 300000 }}
-                            """))
-                        }
-                    case "start":
-                        Task {
-                            try await Task.sleep(nanoseconds: 50_000_000)
-                            await mockWebSocketClient.subject.send(.string("""
-                                {"type": "start_ack", "id": "\(id)"}
-                            """))
-                        }
-                    default:
-                        break
-                    }
-                default:
-                    break
-                }
-            }
-            .store(in: &cancellables)
-
-        // Subscribe and wait for initial subscription
-        let subscription = try await appSyncClient.subscribe(id: id, query: "subscription { onTest { id } }")
-            .sink { event in
-                if case .subscribed = event {
-                    let count = subscribedCount.increment()
-                    if count == 1 {
-                        initialSubscribed.fulfill()
-                    } else {
-                        resubscribedAfterDrop.fulfill()
-                    }
-                }
-            }
-        cancellables.insert(subscription)
-
-        await fulfillment(of: [initialSubscribed], timeout: 5)
-
-        // Simulate connection drop (as happens when iOS defuncts the socket)
-        await mockWebSocketClient.subject.send(.disconnected(.invalid, nil))
-
-        // Simulate WebSocket reconnecting (auto-retry brings it back)
-        try await Task.sleep(nanoseconds: 100_000_000)
-        await mockWebSocketClient.subject.send(.connected)
-
-        await fulfillment(of: [resubscribedAfterDrop], timeout: 10)
-        withExtendedLifetime(cancellables) { }
-    }
-
-    /// Verifies that multiple concurrent subscriptions all recover after a
-    /// connection drop. This reproduces the GH-4220 scenario where 6 workers
-    /// all get stuck in "connecting" because the reconnection/resubscription
-    /// path doesn't handle multiple subscriptions correctly.
-    ///
-    /// - Given:
-    ///    - An AppSyncRealTimeClient with 6 active subscriptions.
-    /// - When:
-    ///    - The WebSocket emits `.disconnected` (connection drop).
-    ///    - Followed by `.connected` (reconnection).
-    ///    - Server responds with `connection_ack` and `start_ack` for each.
-    /// - Then:
-    ///    - All 6 subscriptions receive a second `.subscribed` event.
-    func testMultipleSubscriptions_afterConnectionDrop_shouldAllResubscribe() async throws {
-        var cancellables = Set<AnyCancellable>()
-        let mockWebSocketClient = MockWebSocketClient()
-        let mockAppSyncRequestInterceptor = MockAppSyncRequestInterceptor()
-        let appSyncClient = AppSyncRealTimeClient(
-            endpoint: URL(string: "https://example.com")!,
-            requestInterceptor: mockAppSyncRequestInterceptor,
-            webSocketClient: mockWebSocketClient
-        )
-
-        let subscriptionCount = 6
-        let ids = (0..<subscriptionCount).map { _ in UUID().uuidString }
-
-        let allInitiallySubscribed = expectation(description: "All subscriptions initially established")
-        allInitiallySubscribed.expectedFulfillmentCount = subscriptionCount
-
-        let allResubscribed = expectation(description: "All subscriptions re-established after drop")
-        allResubscribed.expectedFulfillmentCount = subscriptionCount
-
-        // Track per-subscription subscribed counts
-        let subscribedCounts = (0..<subscriptionCount).map { _ in AtomicCounter() }
-
-        await mockWebSocketClient.setStateToConnected()
-
-        // Wire up mock to respond to protocol messages
-        await mockWebSocketClient.actionSubject
-            .sink { action in
-                switch action {
-                case .connect:
-                    Task {
-                        await mockWebSocketClient.subject.send(.connected)
-                    }
-                case .write(let message):
-                    guard let response = try? JSONDecoder().decode(
-                        JSONValue.self,
-                        from: message.data(using: .utf8)!
-                    ) else { return }
-
-                    switch response.type?.stringValue {
-                    case "connection_init":
-                        Task {
-                            try await Task.sleep(nanoseconds: 30_000_000)
-                            await mockWebSocketClient.subject.send(.string("""
-                                {"type": "connection_ack", "payload": { "connectionTimeoutMs": 300000 }}
-                            """))
-                        }
-                    case "start":
-                        let startId = response.id?.stringValue ?? ""
-                        Task {
-                            try await Task.sleep(nanoseconds: 30_000_000)
-                            await mockWebSocketClient.subject.send(.string("""
-                                {"type": "start_ack", "id": "\(startId)"}
-                            """))
-                        }
-                    default:
-                        break
-                    }
-                default:
-                    break
-                }
-            }
-            .store(in: &cancellables)
-
-        // Create all subscriptions
-        for (index, id) in ids.enumerated() {
-            let subscription = try await appSyncClient.subscribe(
-                id: id,
-                query: "subscription { onTest { id } }"
-            ).sink { event in
-                if case .subscribed = event {
-                    let count = subscribedCounts[index].increment()
-                    if count == 1 {
-                        allInitiallySubscribed.fulfill()
-                    } else if count == 2 {
-                        allResubscribed.fulfill()
-                    }
-                }
-            }
-            cancellables.insert(subscription)
-        }
-
-        await fulfillment(of: [allInitiallySubscribed], timeout: 10)
-
-        // Simulate connection drop
-        await mockWebSocketClient.subject.send(.disconnected(.invalid, nil))
-
-        // Simulate reconnection
-        try await Task.sleep(nanoseconds: 100_000_000)
-        await mockWebSocketClient.subject.send(.connected)
-
-        await fulfillment(of: [allResubscribed], timeout: 15)
-        withExtendedLifetime(cancellables) { }
-    }
-
-    /// Verifies that rapidly stopping and restarting subscriptions (as happens
-    /// during iOS foreground recovery) does not leave subscriptions orphaned.
-    /// This reproduces the race condition from GH-4220 where unsubscribe is
-    /// called on a subscription still in "subscribing" state.
-    ///
-    /// - Given:
-    ///    - An AppSyncRealTimeClient in connected state.
-    /// - When:
-    ///    - A subscription is created and immediately unsubscribed before
-    ///      the `start_ack` arrives.
-    ///    - A new subscription with the same query is then created.
-    /// - Then:
-    ///    - The second subscription successfully reaches "subscribed" state.
-    ///    - No subscription is left orphaned.
-    func testRapidUnsubscribeResubscribe_shouldNotOrphanSubscription() async throws {
-        var cancellables = Set<AnyCancellable>()
-        let mockWebSocketClient = MockWebSocketClient()
-        let mockAppSyncRequestInterceptor = MockAppSyncRequestInterceptor()
-        let appSyncClient = AppSyncRealTimeClient(
-            endpoint: URL(string: "https://example.com")!,
-            requestInterceptor: mockAppSyncRequestInterceptor,
-            webSocketClient: mockWebSocketClient
-        )
-
-        await mockWebSocketClient.setStateToConnected()
-
-        // Wire up mock — delay start_ack to simulate network latency
-        await mockWebSocketClient.actionSubject
-            .sink { action in
-                switch action {
-                case .connect:
-                    Task {
-                        await mockWebSocketClient.subject.send(.connected)
-                    }
-                case .write(let message):
-                    guard let response = try? JSONDecoder().decode(
-                        JSONValue.self,
-                        from: message.data(using: .utf8)!
-                    ) else { return }
-
-                    switch response.type?.stringValue {
-                    case "connection_init":
-                        Task {
-                            try await Task.sleep(nanoseconds: 30_000_000)
-                            await mockWebSocketClient.subject.send(.string("""
-                                {"type": "connection_ack", "payload": { "connectionTimeoutMs": 300000 }}
-                            """))
-                        }
-                    case "start":
-                        let startId = response.id?.stringValue ?? ""
-                        // Delayed start_ack — simulates network round-trip
-                        Task {
-                            try await Task.sleep(nanoseconds: 200_000_000)
-                            await mockWebSocketClient.subject.send(.string("""
-                                {"type": "start_ack", "id": "\(startId)"}
-                            """))
-                        }
-                    case "stop":
-                        let stopId = response.id?.stringValue ?? ""
-                        Task {
-                            try await Task.sleep(nanoseconds: 30_000_000)
-                            await mockWebSocketClient.subject.send(.string("""
-                                {"type": "complete", "id": "\(stopId)"}
-                            """))
-                        }
-                    default:
-                        break
-                    }
-                default:
-                    break
-                }
-            }
-            .store(in: &cancellables)
-
-        // Connect first
-        Task {
-            try await Task.sleep(nanoseconds: 50_000_000)
-            await mockWebSocketClient.subject.send(.connected)
-            try await Task.sleep(nanoseconds: 50_000_000)
-            await mockWebSocketClient.subject.send(.string("""
-                {"type": "connection_ack", "payload": { "connectionTimeoutMs": 300000 }}
-            """))
-        }
-        try await appSyncClient.connect()
-
-        // Create first subscription and immediately unsubscribe (before start_ack)
-        let firstId = UUID().uuidString
-        let firstSubscription = try await appSyncClient.subscribe(
-            id: firstId,
-            query: "subscription { onTest { id } }"
-        ).sink { _ in }
-        cancellables.insert(firstSubscription)
-
-        // Unsubscribe quickly — before start_ack arrives
-        try await Task.sleep(nanoseconds: 50_000_000)
-        try await appSyncClient.unsubscribe(id: firstId)
-
-        // Create second subscription — this should succeed
-        let secondId = UUID().uuidString
-        let secondSubscribed = expectation(description: "Second subscription should reach subscribed state")
-        secondSubscribed.assertForOverFulfill = false
-
-        let secondSubscription = try await appSyncClient.subscribe(
-            id: secondId,
-            query: "subscription { onTest { id } }"
-        ).sink { event in
-            if case .subscribed = event {
-                secondSubscribed.fulfill()
-            }
-        }
-        cancellables.insert(secondSubscription)
-
-        await fulfillment(of: [secondSubscribed], timeout: 5)
-        withExtendedLifetime(cancellables) { }
-    }
-
     /// Verifies that when a WebSocket emits a `.error(.connectionLost)` event
     /// (as happens when `didCompleteWithError` fires with ECONNABORTED),
     /// all active subscriptions receive the error and the client transitions
@@ -419,7 +96,7 @@ class AppSyncRealTimeClientWriteFailureTests: XCTestCase {
         )
 
         let subscriptionCount = 3
-        let ids = (0..<subscriptionCount).map { _ in UUID().uuidString }
+        let ids = (0 ..< subscriptionCount).map { _ in UUID().uuidString }
 
         let allSubscribed = expectation(description: "All subscriptions established")
         allSubscribed.expectedFulfillmentCount = subscriptionCount
@@ -500,249 +177,6 @@ class AppSyncRealTimeClientWriteFailureTests: XCTestCase {
         await fulfillment(of: [allReceivedError], timeout: 5)
         withExtendedLifetime(cancellables) { }
     }
-
-    /// Regression test for https://github.com/aws-amplify/amplify-swift/issues/4220.
-    /// Verifies that unsubscribing a subscription in `.subscribing` state
-    /// (waiting for connection) properly marks it as unsubscribed, so the
-    /// in-flight Task doesn't proceed to send `.start` after the connection
-    /// establishes.
-    ///
-    /// - Given:
-    ///    - An AppSyncRealTimeClient that is NOT yet connected.
-    ///    - A subscription is created (its deferred Task is waiting for connection).
-    /// - When:
-    ///    - `unsubscribe(id)` is called before the connection establishes.
-    ///    - The connection then establishes.
-    /// - Then:
-    ///    - No `.start` message is sent for the unsubscribed subscription.
-    ///    - The subscription emits `.unsubscribed`.
-    func testUnsubscribe_whileSubscriptionWaitingForConnection_shouldNotSendStart() async throws {
-        var cancellables = Set<AnyCancellable>()
-        let mockWebSocketClient = MockWebSocketClient()
-        let mockAppSyncRequestInterceptor = MockAppSyncRequestInterceptor()
-        let appSyncClient = AppSyncRealTimeClient(
-            endpoint: URL(string: "https://example.com")!,
-            requestInterceptor: mockAppSyncRequestInterceptor,
-            webSocketClient: mockWebSocketClient
-        )
-
-        let id = UUID().uuidString
-
-        let unsubscribedEvent = expectation(description: "Subscription should emit unsubscribed")
-        let noStartSent = expectation(description: "No start message should be sent for this subscription")
-        noStartSent.isInverted = true
-
-        await mockWebSocketClient.setStateToConnected()
-
-        // Track writes to detect if .start is sent for our subscription
-        await mockWebSocketClient.actionSubject
-            .sink { action in
-                if case .write(let message) = action {
-                    if let response = try? JSONDecoder().decode(JSONValue.self, from: message.data(using: .utf8)!),
-                       response.type?.stringValue == "start",
-                       response.id?.stringValue == id {
-                        noStartSent.fulfill()
-                    }
-                }
-            }
-            .store(in: &cancellables)
-
-        // Subscribe — but don't connect yet, so the deferred Task waits
-        let subscription = try await appSyncClient.subscribe(
-            id: id,
-            query: "subscription { onTest { id } }"
-        ).sink { event in
-            if case .unsubscribed = event {
-                unsubscribedEvent.fulfill()
-            }
-        }
-        cancellables.insert(subscription)
-
-        // Unsubscribe immediately — before connection establishes
-        try await Task.sleep(nanoseconds: 30_000_000)
-        try await appSyncClient.unsubscribe(id: id)
-
-        // Now establish the connection — the unsubscribed subscription's Task
-        // should NOT send .start
-        try await Task.sleep(nanoseconds: 50_000_000)
-        await mockWebSocketClient.subject.send(.connected)
-        try await Task.sleep(nanoseconds: 50_000_000)
-        await mockWebSocketClient.subject.send(.string("""
-            {"type": "connection_ack", "payload": { "connectionTimeoutMs": 300000 }}
-        """))
-
-        await fulfillment(of: [unsubscribedEvent], timeout: 3)
-        // Wait to confirm no .start was sent (inverted expectation)
-        await fulfillment(of: [noStartSent], timeout: 2)
-        withExtendedLifetime(cancellables) { }
-    }
-
-    /// Regression test for https://github.com/aws-amplify/amplify-swift/issues/4220.
-    /// Verifies that when a subscription is unsubscribed while in `.subscribing`
-    /// state, a subsequent subscribe with the same ID works correctly.
-    ///
-    /// - Given:
-    ///    - An AppSyncRealTimeClient in connected state.
-    ///    - A subscription that was unsubscribed while still in `.subscribing` state.
-    /// - When:
-    ///    - A new subscription with the same query is created.
-    /// - Then:
-    ///    - The new subscription successfully reaches `.subscribed` state.
-    func testSubscribe_afterUnsubscribeDuringSubscribing_shouldSucceed() async throws {
-        var cancellables = Set<AnyCancellable>()
-        let mockWebSocketClient = MockWebSocketClient()
-        let mockAppSyncRequestInterceptor = MockAppSyncRequestInterceptor()
-        let appSyncClient = AppSyncRealTimeClient(
-            endpoint: URL(string: "https://example.com")!,
-            requestInterceptor: mockAppSyncRequestInterceptor,
-            webSocketClient: mockWebSocketClient
-        )
-
-        await mockWebSocketClient.setStateToConnected()
-
-        // Wire up mock to respond to protocol messages
-        await mockWebSocketClient.actionSubject
-            .sink { action in
-                if case .write(let message) = action {
-                    guard let response = try? JSONDecoder().decode(
-                        JSONValue.self,
-                        from: message.data(using: .utf8)!
-                    ) else { return }
-
-                    switch response.type?.stringValue {
-                    case "connection_init":
-                        Task {
-                            try await Task.sleep(nanoseconds: 30_000_000)
-                            await mockWebSocketClient.subject.send(.string("""
-                                {"type": "connection_ack", "payload": { "connectionTimeoutMs": 300000 }}
-                            """))
-                        }
-                    case "start":
-                        let startId = response.id?.stringValue ?? ""
-                        // Delayed start_ack
-                        Task {
-                            try await Task.sleep(nanoseconds: 200_000_000)
-                            await mockWebSocketClient.subject.send(.string("""
-                                {"type": "start_ack", "id": "\(startId)"}
-                            """))
-                        }
-                    default:
-                        break
-                    }
-                }
-            }
-            .store(in: &cancellables)
-
-        // Connect first
-        Task {
-            try await Task.sleep(nanoseconds: 50_000_000)
-            await mockWebSocketClient.subject.send(.connected)
-            try await Task.sleep(nanoseconds: 50_000_000)
-            await mockWebSocketClient.subject.send(.string("""
-                {"type": "connection_ack", "payload": { "connectionTimeoutMs": 300000 }}
-            """))
-        }
-        try await appSyncClient.connect()
-
-        // Create first subscription
-        let firstId = UUID().uuidString
-        let firstSub = try await appSyncClient.subscribe(
-            id: firstId,
-            query: "subscription { onTest { id } }"
-        ).sink { _ in }
-        cancellables.insert(firstSub)
-
-        // Unsubscribe before start_ack arrives (subscription is in .subscribing)
-        try await Task.sleep(nanoseconds: 50_000_000)
-        try await appSyncClient.unsubscribe(id: firstId)
-
-        // Now create a new subscription — should work fine
-        let secondId = UUID().uuidString
-        let secondSubscribed = expectation(description: "New subscription should succeed")
-        secondSubscribed.assertForOverFulfill = false
-
-        let secondSub = try await appSyncClient.subscribe(
-            id: secondId,
-            query: "subscription { onTest { id } }"
-        ).sink { event in
-            if case .subscribed = event {
-                secondSubscribed.fulfill()
-            }
-        }
-        cancellables.insert(secondSub)
-
-        await fulfillment(of: [secondSubscribed], timeout: 5)
-        withExtendedLifetime(cancellables) { }
-    }
-
-    /// Regression test for https://github.com/aws-amplify/amplify-swift/issues/4220.
-    /// Verifies that the deferred Task in subscribe() checks if the subscription
-    /// was removed from the dict before sending .start. This prevents orphaned
-    /// Tasks from sending requests for subscriptions that were already cleaned up.
-    ///
-    /// - Given:
-    ///    - An AppSyncRealTimeClient not yet connected.
-    ///    - Multiple subscriptions created (deferred Tasks waiting for connection).
-    /// - When:
-    ///    - All subscriptions are unsubscribed before connection establishes.
-    ///    - Connection then establishes.
-    /// - Then:
-    ///    - No .start messages are sent for any of the unsubscribed subscriptions.
-    func testUnsubscribeAll_beforeConnectionEstablishes_shouldNotSendAnyStarts() async throws {
-        var cancellables = Set<AnyCancellable>()
-        let mockWebSocketClient = MockWebSocketClient()
-        let mockAppSyncRequestInterceptor = MockAppSyncRequestInterceptor()
-        let appSyncClient = AppSyncRealTimeClient(
-            endpoint: URL(string: "https://example.com")!,
-            requestInterceptor: mockAppSyncRequestInterceptor,
-            webSocketClient: mockWebSocketClient
-        )
-
-        let ids = (0..<3).map { _ in UUID().uuidString }
-
-        let noStartsSent = expectation(description: "No start messages should be sent")
-        noStartsSent.isInverted = true
-
-        await mockWebSocketClient.setStateToConnected()
-
-        await mockWebSocketClient.actionSubject
-            .sink { action in
-                if case .write(let message) = action {
-                    if let response = try? JSONDecoder().decode(JSONValue.self, from: message.data(using: .utf8)!),
-                       response.type?.stringValue == "start" {
-                        noStartsSent.fulfill()
-                    }
-                }
-            }
-            .store(in: &cancellables)
-
-        // Create subscriptions — deferred Tasks will wait for connection
-        for id in ids {
-            let sub = try await appSyncClient.subscribe(
-                id: id,
-                query: "subscription { onTest { id } }"
-            ).sink { _ in }
-            cancellables.insert(sub)
-        }
-
-        // Unsubscribe all before connection
-        try await Task.sleep(nanoseconds: 30_000_000)
-        for id in ids {
-            try await appSyncClient.unsubscribe(id: id)
-        }
-
-        // Now establish connection
-        try await Task.sleep(nanoseconds: 50_000_000)
-        await mockWebSocketClient.subject.send(.connected)
-        try await Task.sleep(nanoseconds: 50_000_000)
-        await mockWebSocketClient.subject.send(.string("""
-            {"type": "connection_ack", "payload": { "connectionTimeoutMs": 300000 }}
-        """))
-
-        // Wait to confirm no .start was sent
-        await fulfillment(of: [noStartsSent], timeout: 3)
-        withExtendedLifetime(cancellables) { }
-    }
 }
 
 // MARK: - Test Helpers
@@ -788,25 +222,5 @@ private actor MockWriteFailingWebSocketClient: AppSyncWebSocketClientProtocol {
 
     func setShouldFailWrites(_ shouldFail: Bool) {
         shouldFailWrites = shouldFail
-    }
-}
-
-/// Thread-safe counter for tracking subscription events across concurrent tasks.
-private final class AtomicCounter: @unchecked Sendable {
-    private var value: Int = 0
-    private let lock = NSLock()
-
-    @discardableResult
-    func increment() -> Int {
-        lock.lock()
-        defer { lock.unlock() }
-        value += 1
-        return value
-    }
-
-    var current: Int {
-        lock.lock()
-        defer { lock.unlock() }
-        return value
     }
 }
